Create Java 1D Array.java
Input Format
_________________________________________
The first line contains a single integer, n, denoting the size of the array.
Each line i of the n subsequent lines contains a single integer denoting the value of element ai.
_________________________________________________
Output Format
You are not responsible for printing any output to stdout. Locked code in the editor loops through array a and prints each sequential element on a new line.